renamed extract to incidents_extract
