+++ pipelines/incidents-data/test_extract_incidents.py
@@ -5,8 +5,8 @@ import xml.etree.ElementTree as ET
 
 import pytest
 
-from .extract import (is_paddington_to_bristol,
-                      extract_relevant_data, parse_xml, get_incident_data)
+from incidents_extract import (is_paddington_to_bristol,
+                               extract_relevant_data, parse_xml, get_incident_data)
 
 # Test is_paddington_to_bristol
